Refactor artist pin functionality to use new module path
- Updated import path for toggleArtistPin function to reflect new location in accountArtistIds directory.
- Introduced toggleArtistPin function to manage artist pin status in the account_artist_ids table.
